finalize tacotron2

+++ tensorflow_tts/models/tacotron2.py
@@ -1,7 +1,17 @@
 # -*- coding: utf-8 -*-
-
-# Copyright 2020 Minh Nguyen (@dathudeptrai) - Eren Gölge (@erogol)
-#  MIT License (https://opensource.org/licenses/MIT)
+# Copyright 2020 The Tacotron-2 Authors, Minh Nguyen (@dathudeptrai) and Eren Gölge (@erogol)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 """Tacotron-2 Modules."""
 
@@ -650,7 +660,7 @@ class TFTacotron2(tf.keras.Model):
         input_ids = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9]])
         input_lengths = np.array([9])
         speaker_ids = np.array([0])
-        mel_outputs = np.random.normal(size=(1, 50, 80))
+        mel_outputs = np.random.normal(size=(1, 50, 80)).astype(np.float32)
         mel_lengths = np.array([50])
         self(input_ids, input_lengths, speaker_ids, mel_outputs, mel_lengths, 10, training=True)
